ads1118: fix configuration error if no ads1118 devices are configured
It doesn't appear possible to have both the "chip" support and the
temperature sensor support in the same file.  Splitting these out
and also laying the groundwork to support other types of
thermocouples via ads1118.
