+++ config/printer-makerbot-replicator2-2012.cfg
@@ -55,7 +55,7 @@ rotation_distance: 33.0168
 nozzle_diameter: 0.400
 filament_diameter: 1.750
 heater_pin: PE5
-sensor_type: ads1118_thermocouple
+sensor_type: ads1118_typek
 sensor_pin: ads1118:pin_0
 min_temp: 0
 max_temp: 260
@@ -69,7 +69,7 @@ max_temp: 260
 #nozzle_diameter: 0.400
 #filament_diameter: 1.750
 #heater_pin: PE3
-#sensor_type: ads1118_thermocouple
+#sensor_type: ads1118_typek
 #sensor_pin: ads1118:pin_1
 #min_temp: 0
 #max_temp: 260
