+++ klippy/extras/temperature_sensors.cfg
@@ -25,7 +25,7 @@
 [spi_temperature]
 
 # Load "ADS1118" sensors
-[ads1118]
+[ads1118_thermocouple]
 
 # Load "temperature_host" sensor
 [temperature_host]
